Check description for full summary of changes
[tail]
* (2014-08-22 19:49): Fixed AJL (in manually merging code yesterday (to
fix the LDESTROY cascade / charlie-foxtrot), I accidentally caused a
"set list set list [...]" ... which is obviously wrong :p)
* (2014-08-22 21:49): Removed LKILL (recinding CHANGELOG #14; as with
#24, reverting changes made in CHANGELOGs #12 and #14, due to flag
processing issues. LDESTROY was upgraded and brought in from SB6;
LREMOVE was also brought in as-is and not renamed as LKILL as originally
intended)
* (2014-08-22 23:06): Added DATA [ARRAY] CONCAT (to compliment DATA
[ARRAY] LAPPEND)
* (2014-08-22 23:33): Added LCOMMON (ported from SB6)

[core]
* (2014-08-22 20:05): Fixed core loader (sb7.tcl: the RETURN at the end
will kill the full-size script)

[access - set]
* (2014-08-22 20:20): Updated SET, INFO, & WHOIS (added RANK option
(code was already there but not implemented))

[access]
* (2014-08-22 20:57): Fixed WHOIS (had to undo the UNIQUEMATCH code
(which spawned the changes that caused the Charlie-FoxTrot last week);
replaced with appropriate LDESTROY code)

[help]
* (2014-08-22 22:05): Fixed HELP (typoes in help text)

[botmon]
* (2014-08-23 00:17): Added BOTMON IGNORE MODE (allows the dynamic
processing of bots that are on the botnet but not accounted-for through
BOTMON's various settings)
* (2014-08-23 00:21): Upgraded BOTMON STATUS (showing IGNORE MODE)
* (2014-08-23 00:22): Upgraded BOTMON (properly shows IGNORE MODE effect
on both possibilities: all bots present, or, some bots missing. Why
these are coded as completely different sections is beyond me)