StormBot V7 ChangeLog: 2014-08-29 01:15:24 +0000
StormBot V7 ChangeLog: 2014-08-29 01:15:24 +0000

[sb7_tail.tcl]
 *
2014-08-28 18:10:49 +0000: #0132: Fixed FLAGS -PARAMS (the scan-through
for flags always tripped an end-of-phrase and went on to process the
apparent flag. Now, will check if the alleged flag is really a valid
flag before breaking anything and re-processing)
 * 2014-08-28 17:50:22
+0000: #0130: Fixed SB7:SETUP & CONFIG -GLOBAL SHORTCUT ADD (forced all
shorts to be added as lowercase; in previous SB versions, case wasn't
considered. This actually caused the potential of DUPLICATE(!) binds
being created: each of a different case. Found by
Demonicpagan)

[config.sb7]
 * 2014-08-28 17:50:22 +0000: #0130: Fixed
SB7:SETUP & CONFIG -GLOBAL SHORTCUT ADD (forced all shorts to be added
as lowercase; in previous SB versions, case wasn't considered. This
actually caused the potential of DUPLICATE(!) binds being created: each
of a different case. Found by Demonicpagan)

[maintenance.sb7]
* (2014-08-26 23:25): Added VERSION into MAINTENANCE bead

Fixed some
issues with binds, cosmetic for the git log generator, and a global
problem with the enhanced flags parser (was clipping text when a
potential flag was found amongst some text, without checking if it WAS a
flag or not first).

Any questions, see us at chat.bothouse.net in
#BotHouse

~ Mai Mizuno, StormBot.TCL coder
Signed-off-by: Dustin <demonicpagan@gmail.com>